requirements.txt
openai==1.13.3
python-pptx==0.6.22
python-dotenv==1.0.0
streamlit==1.32.2